Update MQ2Rez quite a bit of changes.
Reduce down to a single file. 
Correct "extern" in cpp for autorezaccept.
remove unsigned __int64 and unsigned long long in favor of uint64_t
Now in SetGameState if you're at the character select and the plugin is "initialized" it will set it to false, so it will load character specific settings. Applies to switching characters error.
various if statements have had an extra line placed between them for readability.
reduced indentation in a few functions. 
changed from CSidlScreenWnd* to CXWnd* for access to IsVisible(). Saw no reason to cast the value returned from FindMQ2Window since CXWnd* already had access to this function.
compress safemode checks. 
Re-organized "ShouldTakeRez() to check if release to bind is true first, instead of trying to read all the window information first. 
changed BOOL to bool and PCHAR to char* 
provided more detail on valid percentages to choose from when doing /rez pct ## and providing an invalid value.
in /rez setcommand, if no value is found for the 2nd arg, it's set to "DISABLED". 
Added a "/rez silent" option. when set to on will prevent the plugin from outputting spam to the MQ2 window when it accepts a rez.
Added missing "safemode" information from DisplayHelp() function. 
removed the  for loop for the atob function, since we're doing a !_stricmp, it's already case insensitive.

Some of these changes I made a while back. I forgot I even made the changes at some point. So while the date is incorrect for this commit. It's correct for when it actually happened.