MQ2Rez - Reorganize to remove prototypes, adjust SetGameState
in SetGameState - Remove InGame() verification and only set the INI if !Initialized. When Initilizing, set the INI, Do the output messages, and the Update Settings from the INI. 
Remove the above setup from the OnPulse

Reorganize the functions so that they are in an order that doesn't require prototypes anymore. 
Update version float to to be in line with comments.